* Fixed a few compile errors introduced in the last commit * Added Settings.SEND_AGENTUPDATES and Settings.AGENT_UPDATE_INTERVAL, AgentUpdate packets are automatically sent by default now
git-svn-id: http://libopenmetaverse.googlecode.com/svn/trunk@1006 52acb1d6-8a22-11de-b505-999d5b087335

+++ libsecondlife-cs/BitPack.cs
@@ -121,8 +121,8 @@ namespace libsecondlife
 
                 while (count > 0)
                 {
-                    if (data[curBytePos] & (0x80 >> curBitPos) != 0)
-                        Data[bytePos] |= 0x80 >> bitPos++;
+                    if ((data[curBytePos] & (0x80 >> curBitPos)) != 0)
+                        Data[bytePos] |= (byte)(0x80 >> bitPos++);
 
                     --count;
                     ++curBitPos;
